refactor: extract navigation positioning logic to separate module
